organized the files, set up a project manager page, and added some documentation to psudocode.md. much preperation for v5.0
