+++ project.md
@@ -1,0 +1,8 @@
+# Minesweeper:
+
+## TODO:
+- rewrite the code for v5.0
+- make a tool that allows users to click on tiles and have it automatically assign the color value to the custom color.
+- allow users to save custom profiles
+- clean up GUI
+- make instructions for calibration more clear
