+++ v4/Settings.json
@@ -1,0 +1,56 @@
+{
+    "GoogleEasy":{
+        "originPoint":[0, 225],
+        "boardWidth":10,
+        "boardHeight":8,
+        "squareSize":45,
+        "cOne":[25, 118, 210],
+        "cTwo":[56, 142, 60],
+        "cThree":[211, 47, 47],
+        "cFour":[137, 52, 162],
+        "cFive":[255, 143, 0],
+        "cSix":[0, 151, 167],
+        "cSeven":[109, 100, 91],
+        "cGreen":[162, 209, 73],
+        "cFlag":[242, 54, 7],
+        "defaultOffset":[22, -32],
+        "blueSpace":[850, 500],
+        "minSimilarity":10
+    },
+    "GoogleMedium":{
+        "originPoint":[0, 210],
+        "boardWidth":18,
+        "boardHeight":14,
+        "squareSize":30,
+        "cOne":[25, 118, 210],
+        "cTwo":[56, 142, 60],
+        "cThree":[211, 47, 47],
+        "cFour":[123, 31, 162],
+        "cFive":[255, 143, 0],
+        "cSix":[0, 151, 167],
+        "cSeven":[109, 100, 91],
+        "cGreen":[170, 215, 81],
+        "cFlag":[242, 54, 7],
+        "defaultOffset":[16, -21],
+        "blueSpace":[850, 500],
+        "minSimilarity":10
+    },
+    "GoogleHard":{
+        "originPoint":[0, 205],
+        "boardWidth":24,
+        "boardHeight":20,
+        "squareSize":25,
+        "cOne":[25, 118, 210],
+        "cTwo":[111, 156, 92],
+        "cThree":[218, 105, 91],
+        "cFour":[136, 52, 161],
+        "cFive":[249, 150, 24],
+        "cSix":[0, 151, 167],
+        "cSeven":[109, 100, 91],
+        "cGreen":[170, 215, 81],
+        "cFlag":[242, 54, 7],
+        "defaultOffset":[12, -17],
+        "blueSpace":[850, 500],
+        "minSimilarity":10
+    }
+}
